Restructure test folders to match src folder
